requirements.txt
fire
opencv-python
pillow
piexif
requests
rich
diffusers
transformers
xformers
accelerate
opencv-contrib-python
controlnet_aux
matplotlib
ipykernel
mediapipe
moviepy
more-itertools